move the atom_in.h and it's test file atom_in_test.cpp to the source/module_base. atom_in.h contain atom symbol map like H, 1
