+++ source/module_base/test/CMakeLists.txt
@@ -5,6 +5,10 @@ AddTest(
   SOURCES blas_connector_test.cpp
 )
 AddTest(
+  TARGET surchem_atom_in
+  SOURCES atom_in_test.cpp
+)
+AddTest(
   TARGET base_blacs_connector
   LIBS ${math_libs}
   SOURCES blacs_connector_test.cpp
